Change uint32 IDs to uint64 in database operations
Converted various ID fields and parameters from uint32 to uint64 to accommodate larger identifier values. This affects user, role, and permission IDs across different modules, ensuring consistency and support for larger IDs.

+++ authentication/service.go
@@ -18,8 +18,10 @@ import (
 	"goflare.io/auth/user"
 )
 
+// _ is used to ensure that *service implements the Service interface at compile time.
 var _ Service = (*service)(nil)
 
+// Service represents the set of methods for user authentication and authorization.
 type Service interface {
 	// Login logs in a user with email and password.
 	Login(ctx context.Context, email, password string) (*models.PASETOToken, error)
@@ -28,11 +30,13 @@ type Service interface {
 	// Register registers a new user with username, password, email, and phone.
 	Register(ctx context.Context, username, password, email, phone string) (*models.PASETOToken, error)
 	// ValidateToken validates a token.
-	ValidateToken(token string) (uint32, error)
+	ValidateToken(token string) (uint64, error)
 	// CheckPermission checks if a user has a permission for a resource and action.
-	CheckPermission(ctx context.Context, userID uint32, resource enum.ResourceType, action enum.ActionType) (bool, error)
+	CheckPermission(ctx context.Context, userID uint64, resource enum.ResourceType, action enum.ActionType) (bool, error)
 }
 
+// service represents the core service implementation for user management and authorization.
+// It integrates various components such as the user repository, token manager, policy enforcer, and logger.
 type service struct {
 	userStore    user.Repository
 	tokenManager token.Manager
@@ -40,6 +44,7 @@ type service struct {
 	logger       *zap.Logger
 }
 
+// NewService creates a new instance of Service with a provided user repository, configuration, enforcer, and logger.
 func NewService(userStore user.Repository, config *config.Config, enforcer *casbin.Enforcer, logger *zap.Logger) Service {
 
 	tokenManager := token.NewPasetoManager(config.Paseto.PublicSecretKey, config.Paseto.PrivateSecretKey, config.Paseto.TokenExpirationTime)
@@ -51,6 +56,8 @@ func NewService(userStore user.Repository, config *config.Config, enforcer *casb
 	}
 }
 
+// Login authenticates a user with the provided email and password,
+// returning a PASETO token upon successful authentication.
 func (s *service) Login(ctx context.Context, email, password string) (*models.PASETOToken, error) {
 	s.logger.Info("login user", zap.String("email", email))
 	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
@@ -71,11 +78,13 @@ func (s *service) Login(ctx context.Context, email, password string) (*models.PA
 	return s.tokenManager.GenerateToken(user.ID)
 }
 
+// Logout revokes the provided token, effectively logging the user out.
 func (s *service) Logout(ctx context.Context, token string) error {
 	s.logger.Info("logout user", zap.String("token", token))
 	return s.tokenManager.RevokeToken(token)
 }
 
+// Register registers a new user, hashes the password, stores the user, and generates a PASETO token.
 func (s *service) Register(ctx context.Context, username, password, email, phone string) (*models.PASETOToken, error) {
 
 	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
@@ -102,13 +111,15 @@ func (s *service) Register(ctx context.Context, username, password, email, phone
 	return s.tokenManager.GenerateToken(user.ID)
 }
 
-func (s *service) ValidateToken(token string) (uint32, error) {
+// ValidateToken validates the given token and returns the associated user ID if valid, otherwise an error.
+func (s *service) ValidateToken(token string) (uint64, error) {
 	s.logger.Info("validate token", zap.String("token", token))
 	return s.tokenManager.ValidateToken(token)
 }
 
-func (s *service) CheckPermission(ctx context.Context, userID uint32, resource enum.ResourceType, action enum.ActionType) (bool, error) {
-	s.logger.Info("check permission", zap.Uint32("userID", userID), zap.Any("resource", resource), zap.Any("action", action))
+// CheckPermission verifies if the user has permission to perform a specific action on a resource.
+func (s *service) CheckPermission(ctx context.Context, userID uint64, resource enum.ResourceType, action enum.ActionType) (bool, error) {
+	s.logger.Info("check permission", zap.Uint64("userID", userID), zap.Any("resource", resource), zap.Any("action", action))
 	ctx, cancel := context.WithTimeout(ctx, 5*time.Second)
 	defer cancel()
