+++ permission/service.go
@@ -9,8 +9,8 @@ import (
 // Service is the interface for the permission service.
 type Service interface {
 	Create(ctx context.Context, permission *models.Permission) error
-	GetByID(ctx context.Context, id uint32) (*models.Permission, error)
-	Delete(ctx context.Context, id uint32) error
+	GetByID(ctx context.Context, id uint64) (*models.Permission, error)
+	Delete(ctx context.Context, id uint64) error
 }
 
 // service is the implementation of the Service interface.
@@ -34,13 +34,13 @@ func (s *service) Create(ctx context.Context, permission *models.Permission) err
 }
 
 // GetByID gets a permission by ID.
-func (s *service) GetByID(ctx context.Context, id uint32) (*models.Permission, error) {
+func (s *service) GetByID(ctx context.Context, id uint64) (*models.Permission, error) {
 
 	return s.repo.GetByID(ctx, id)
 }
 
 // Delete deletes a permission by ID.
-func (s *service) Delete(ctx context.Context, id uint32) error {
+func (s *service) Delete(ctx context.Context, id uint64) error {
 
 	return s.repo.Delete(ctx, id)
 }
